requirements.txt
torch>=2.2.0
transformers>=4.45.0
accelerate>=0.33.0
bitsandbytes>=0.43.0; platform_system != "Darwin"
sentencepiece>=0.2.0
protobuf>=4.25.0
datasets>=2.20.0
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=3.0.0
simhash>=2.1.2
matplotlib>=3.8.0
seaborn>=0.13.0
openai>=1.50.0
tqdm>=4.66.0
pyyaml>=6.0
psutil>=5.9.0
